+++ CHANGELOG.md
@@ -6,6 +6,12 @@ Database schema version: 2
 
 - added database migration to create `account_public_key_bindings` and inserts data for all account public key mappings into this table
 
+- Added `PublicKeyBindingInfo` type that represents public key info of accounts access structure.
+- Added `PublicKeyBindingInfo` into `TransactionLogData`.
+- Added `insert_key_bindings` and `delete_key_bindings` feilds into `PreparedStatements`.
+- The DB task is extended to save account's address and public key binding.
+- The Node task is extended to monitor for public key update and new account creation transactions.
+
 ## [0.14.0] - 2025-08-07
 
 ### Changed